refactoring generateTopoGraph
