+++ src/task/npmTask.ts
@@ -2,12 +2,12 @@ import { Config } from "../types/Config";
 import { controller, signal } from "./abortSignal";
 import { findNpmClient } from "../workspace/findNpmClient";
 import { PackageInfo } from "workspace-tools";
+import { RunContext } from "../types/RunContext";
 import { spawn } from "child_process";
-import { taskWrapper } from "./taskWrapper";
 import { taskLogger, NpmLogWritable } from "../logger";
+import { taskWrapper } from "./taskWrapper";
 import path from "path";
 import PQueue from "p-queue";
-import { RunContext } from "../types/RunContext";
 
 function wait(time: number): Promise<void> {
   return new Promise((resolve, reject) => {
